+++ CHANGELOG.md
@@ -4,8 +4,8 @@ This project follows [Semantic Versioning](http://semver.org/).
 
 ## [41.3.0] 2026-04-29
 ### Changed
-- LForms.Util.validateQuestionnaireOnFHIRServer() now returns the
-  whole OperationOutcome object returned from server.
+- LForms.Util.validateQuestionnaireOnFHIRServer() returned Promise
+  now resolves to an OperationOutcome resource from the $validate operation.
 
 ## [41.2.0] 2026-04-20
 ### Added